pyproject.toml
[tool.poetry]
name = "predicting-cyclodextrin-bioavailability"
version = "0.1.0"
description = "Machine-learning toolkit for predicting how cyclodextrins modulate the bioavailability of small-molecule drugs."
authors = ["Macrume <jakub.jarzabek49@gmail.com>"]
readme = "README.md"
packages = [{ include = "predicting-cyclodextrin-bioavailability" }]


[tool.poetry.dependencies]
python = ">=3.11,<3.13"
ipykernel = "^6.29.5"
pandas = "^2.2.3"
numpy = "^2.2.5"
matplotlib = "^3.10.1"
seaborn = "^0.13.2"
scikit-learn = "^1.6.1"
xgboost = "^3.0.0"
pyarrow = "^19.0.1"
optuna = "^4.3.0"
shap = "^0.47.2"
ipywidgets = "^8.1.6"
lightgbm = "^4.6.0"
catboost = "^1.2.8"
optuna-integration = {extras = ["xgboost"], version = "^4.3.0"}
tabulate = "^0.9.0"
hmbfr = {path = "../hmbfr", develop = true}
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
